docs: reversed decision to include params in workflow_dispatch
When you add input parameters to the workflow_dispatch condition, the
only way to provide default values (to get values into the forms UI
when invoking the workflow manually on GitHub) is to write the values
literally for `default:`. There is no way to get the values from the
`env` block using, e.g., `${{env.files}}`, because GitHub Actions
provides no mechanism by which you can do that. This means you end up
having to duplicate the default parameter values in the `env` block
and again in the `on > workflow_dispatch > inputs` block.

I'm worried this will lead to difficult-to-detect errors and thus be
less maintainable, both for us and for users. I also suspect that
users probably will _usually_ not need to change the values when
invoking the workflow manually on GitHub. Thus, I think it's better to
omit them from the sample workflow.

The workflow in `.github/workflows/bad-link-reporter.yml` still has
them.