fix: revise push vspull, also cron, also job name
After using this for a repository, I think I'm wrong about not testing
on pushes and testing on PRs. You run into a problem with editing and
doing PRs on GitHub: you want the test to be upon saving the file,
which is a push, not a PR. And having both push and pull_request means
that the checks are run twice when someone opens a PR.

If this gets run on pushes, then it doesn't make sense to run it with
cron on a daily basis. I changed that to a once-a-week schedule.